+++ src/index.js
@@ -33,7 +33,35 @@ function mergeStrategy(rules = {}) {
         },
         customizeObject: (a, b, key) => {
           if (rules[key]) {
-            // No support for recursive checks yet - are those even needed?
+            return rules[key] === 'prepend' && lodashMerge({}, b, a, joinArrays());
+          }
+        }
+      })
+    ]));
+  };
+}
+
+function mergeSmartStrategy(rules = {}) {
+  // rules: { <field>: <'append'|'prepend'> }
+  // All default to append but you can override here
+  return function () {
+    return lodashMerge.apply(null, [{}].concat(...arguments).concat([
+      joinArrays({
+        customizeArray: (a, b, key) => {
+          if (isLoader(key.split('.').slice(-1)[0])) {
+            if (rules[key]) {
+              return b.reduce(reduceLoaders, a.slice());
+            }
+
+            return a.reduce(reduceLoaders, b.slice());
+          }
+
+          if (rules[key]) {
+            return rules[key] === 'prepend' && b.concat(a);
+          }
+        },
+        customizeObject: (a, b, key) => {
+          if (rules[key]) {
             return rules[key] === 'prepend' && lodashMerge({}, b, a, joinArrays());
           }
         }
@@ -49,3 +77,4 @@ function isLoader(key) {
 module.exports = merge;
 module.exports.smart = mergeSmart;
 module.exports.strategy = mergeStrategy;
+module.exports.smartStrategy = mergeSmartStrategy;